* dda/pr/7709:
  Having not used astyle before, I was copying-pasting too literally into vim and not reading the error message. Simple fix, and I'm now an astyle convert.

* dda/pr/7716:
  * Please know this is a bit of a black box and only via the grand oracle google, right hand lackey stack overflow, and various other articles, I think the one compiler flag I added and the suggestions for compiling on OSX are now correct.

+++ src/computer.cpp
@@ -556,7 +556,7 @@ void computer::activate_function(computer_action action)
             for (int y = 0; y < SEEY * MAPSIZE; y++) {
                 for (size_t i = 0; i < g->m.i_at(x, y).size(); i++) {
                     if (g->m.i_at(x, y)[i].is_bionic()) {
-                        if ((ssize_t)names.size() < TERMY - 8) {
+                        if ((int)names.size() < TERMY - 8) {
                             names.push_back(g->m.i_at(x, y)[i].tname());
                         } else {
                             more++;
